refactor(transfer): decompose generator/transfer.rs into per-concern submodules
Split the 1020-line generator/transfer.rs into a directory:
- mod.rs (re-exports + entry)
- orchestrator.rs (top-level transfer driver)
- transfer_loop.rs (per-file iteration)
- goodbye.rs (goodbye phase)
- stats.rs (per-transfer counters)

Public API preserved via re-exports from mod.rs.
